added scheduler for backup, modified logging, it won't log for admin anymore, implemented backup to all the files each 6 months, generating reports as pdf done.

+++ src/main/java/com/humam/security/group/GroupService.java
@@ -105,12 +105,16 @@ public class GroupService {
     }
 
     public String removeMember(String token, Integer gid, Integer memberId){
-        // TODO don't remove group owner >:-(
+
         boolean owner = this.isGroupOwner(token, gid);
         if(owner){
-            GroupMember groupMember = groupMemberRepository.findRecordByGroupIdAndUserId(gid, memberId);
-            groupMemberRepository.delete(groupMember);
-            return "removed successfully";
+            if(gid != memberId) {
+                GroupMember groupMember = groupMemberRepository.findRecordByGroupIdAndUserId(gid, memberId);
+                groupMemberRepository.delete(groupMember);
+                return "removed successfully";
+            } else {
+                throw new IllegalStateException("can't remove the owner");
+            }
         } else {
             throw new IllegalStateException("not group owner");
         }
